backend/requirements.txt
websockets==12.0
asyncio==3.4.3
asyncpg==0.29.0
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.7
aiohttp==3.9.1
backoff==2.2.1
python-binance==1.0.19
circuitbreaker==2.0.0